Added extra comments for clarity, moved exceptions to exceptions folder, added validation for post description length
